LoreBitsBE/requirements.txt
django
djangorestframework
django-cors-headers
rest-framework-simplejwt
